+++ TeamProject/static/js/board/comment/index.js
@@ -1,0 +1,1 @@
+// crud js
