+++ TeamProject/static/js/board/fetch.js
@@ -1,5 +1,5 @@
 import * as LINK from "../config.js"
-import * as MAIN from "./main.js"
+import * as MAIN from "./main.js.js"
 
 //보드 게시판 (개별)조회
 export async function fetch_getBoard(board_id) {
