+++ TeamProject/static/js/board/main.js
@@ -1,6 +1,6 @@
-import * as FETCH from "./fetch.js";
-import * as REND from "./render.js";
-import * as EVENT from "./event.js";
+import * as FETCH from "./fetch.js.js";
+import * as REND from "./render.js.js";
+import * as EVENT from "./event.js.js";
 // POST_PAGE_COUNT는 무한스크롤시 증가하는 페이지 넘버 , post 로드시에 초기화된다.
 let POST_PAGE_COUNT = 1;
 /*
@@ -27,7 +27,7 @@ export const get_htmlObject = (tag, A, B, C) => {
   return object;
 }
 
-//post_title div에 해당하는 board(게시판)정보 조회 및  가공
+//post_title div에 해당하는 board(게시판)정보 조회 및 가공
 export async function loading_post_title(hashValue) { // load_board()
   try {
     const board = await FETCH.fetch_getBoard(hashValue[1]); //보드 정보 서버에서 받아옴
