+++ TeamProject/static/js/board/post/render.js
